Optimization/Latency reduction in the clientside cloud code

+++ client/src/gfx/clouds.h
@@ -2,5 +2,5 @@
 #include "../../../common/src/common.h"
 
 void cloudsRender();
-void cloudsDraw  (int cx, int cy, int cz);
+void cloudsDraw  (u8 cx, u8 cy, u8 cz);
 void cloudsInit  ();
